Make conceptual shift towards Oracle as trying to achieve desired state
A great way to achieve separation of concerns. Instead of having the
Oracle delete, insert etc. according to implicit rules defined in the
Oracles code, we can instead think of the Oracle as trying to close the
gap between the Algorand buffer and the ideal desired state of the
buffer.

Then, the task of the Oracle boils down to identifying how these two
states differ, and how to close the gap as efficiently as possible